Update to part4/bloglistbackend - completed exercises 4.15 through 4.23

+++ part4/bloglistbackend/controllers/blogs.js
@@ -3,13 +3,17 @@ const Blog = require('../models/blog')
 
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
 
 blogsRouter.post('/', async (request, response) => {
   const { title, author, url, likes } = request.body
+  const user = request.user
+  if (!user) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
   if (!title || !url) {
     response.status(400).json({ error: 'Title and URL are required' })
     return
@@ -19,17 +23,29 @@ blogsRouter.post('/', async (request, response) => {
     author,
     url,
     likes: likes === undefined ? 0 : likes, // Default to 0 if likes is not provided
+    user: user._id
   })
-  const result = await blog.save()
-  response.status(201).json(result)
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+  response.status(201).json(savedBlog)
 })
 
 
 blogsRouter.delete('/:id', async (request, response) => {
-  const result = await Blog.findByIdAndDelete(request.params.id)
-  if (!result) {
+  const user = request.user
+  if (!user) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
     return response.status(404).json({ error: 'Blog not found' })
   }
+  // Ensure that only the creator can delete the blog
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(403).json({ error: 'Permission denied: Only the creator can delete this blog' })
+  }
+  await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
